Move static commands out of main
