--- src/main/java/command/Find.java
@@ -1,4 +1,0 @@
-package main.java.command;
-
-public class Find {
-}
