--- src/main/java/command/Mark.java
@@ -1,4 +1,0 @@
-package main.java.command;
-
-public class Mark {
-}
