+++ src/main/java/msrainy/MsrainyException.java
