+++ src/main/java/msrainy/command/Find.java
@@ -1,0 +1,23 @@
+package main.java.msrainy.command;
+
+import main.java.msrainy.command.task.Task;
+
+import java.util.List;
+
+public class Find {
+    public static void find(List<Task> tasks, String keyword) {
+        boolean found = false;
+        if (tasks.isEmpty()) { System.out.println("\tThere are no tasks"); }
+        for (int i = 0; i < tasks.size(); i++) {
+            Task task = tasks.get(i);
+            String description = task.getDescription();
+            if (description.contains(keyword)) {
+                System.out.println("\t" + i + ". " + tasks.get(i));
+                found = true;
+            }
+        }
+        if (!found) {
+            System.out.println("\tThere are no contain this keyword");
+        }
+    }
+}
