+++ src/main/java/msrainy/command/task/Event.java
@@ -1,3 +1,5 @@
+package main.java.msrainy.command.task;
+
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
