+++ src/main/java/msrainy/command/task/Task.java
@@ -1,3 +1,5 @@
+package main.java.msrainy.command.task;
+
 public abstract class Task {
     protected String description;
     protected boolean isDone;
