+++ src/main/java/msrainy/command/task/ToDo.java
@@ -1,3 +1,5 @@
+package main.java.msrainy.command.task;
+
 public class ToDo extends Task {
     public ToDo(String description) {
         super(description);
