+++ src/main/java/msrainy/ui/Welcome.java
@@ -1,4 +1,8 @@
 package main.java.msrainy.ui;
 
 public class Welcome {
+    public static void welcome() {
+        System.out.println("Hello, I am\n"
+                + "=== MS RAINY ===");
+    }
 }
